More refactoring
Moving methods to more appropriate places.
Added MultiExtractor object with apply and createLazy method (the latter was moved from Extractors where it was called lazyMultiExtractor).
