+++ src/main/scala/com/phasmidsoftware/kmldoc/KML.scala
@@ -64,7 +64,7 @@ trait Feature extends KmlObject
 object Feature extends Extractors with Renderers {
     private val labels: Seq[String] = Seq("Folder", "Document", "Placemark")
     implicit val multiExtractor: MultiExtractor[Seq[Feature]] =
-        lazyMultiExtractor(multiExtractor3[Feature, (Folder, Document, Placemark), Folder, Document, Placemark]((f, d, p) => (f, d, p), labels) ^^ "multiExtractorFeature")
+        MultiExtractor.createLazy(multiExtractor3[Feature, (Folder, Document, Placemark), Folder, Document, Placemark]((f, d, p) => (f, d, p), labels) ^^ "multiExtractorFeature")
     implicit val seqExtractor: Extractor[Seq[Feature]] = seqExtractorByTag("features", labels)
     implicit val renderer: Renderable[Feature] = new Renderers {}.lazyRenderer(rendererSuper2[Feature, Placemark, Container] ^^ "rendererFeature")
     implicit val seqRenderer: Renderable[Seq[Feature]] = sequenceRenderer[Feature] ^^ "rendererFeatures"
